+++ CHANGELOG.md
@@ -7,6 +7,10 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [Unreleased]
 
+## 3.5.1
+### Fixed
+- Fixed an issue with the release CI
+
 ## 3.5.0
 
 ### Added